Ml_ops/01/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.1.2+cpu
torchvision==0.16.2+cpu
timm==0.9.12
hydra-core==1.3.0